extract Banners
