+++ platform/jewel/samples/standalone/src/main/kotlin/org/jetbrains/jewel/samples/standalone/viewmodel/ComponentsViewModel.kt
@@ -6,8 +6,8 @@ import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshots.SnapshotStateList
+import org.jetbrains.jewel.samples.showcase.components.Banners
 import org.jetbrains.jewel.samples.standalone.StandaloneSampleIcons
-import org.jetbrains.jewel.samples.standalone.view.component.Banners
 import org.jetbrains.jewel.samples.standalone.view.component.Borders
 import org.jetbrains.jewel.samples.standalone.view.component.Buttons
 import org.jetbrains.jewel.samples.standalone.view.component.Checkboxes
